refactor: isolate image routing and protocol message assembly
