+++ internal/agent/imageinput.go
@@ -8,40 +8,38 @@ import (
 	"reasonix/internal/provider"
 )
 
-func newImageInput(cfg *imageinput.Config) *imageinput.Service {
-	if cfg == nil {
-		return nil
-	}
-	return imageinput.New(*cfg)
+type agentImageInput struct {
+	service *imageinput.Service
+	native  bool
 }
-func (a *Agent) ImageInput() *imageinput.Service { return a.imageInput }
-func (a *Agent) processToolImages(ctx context.Context, text string, images []string) (string, *provider.VisionSummary) {
+type imageResult struct {
+	text       string
+	summary    *provider.VisionSummary
+	diagnostic error
+}
+
+func newImageInput(cfg *imageinput.Config, p provider.Provider) agentImageInput {
+	if cfg == nil {
+		return agentImageInput{native: supportsNativeImages(p)}
+	}
+	return agentImageInput{service: imageinput.New(*cfg), native: supportsNativeImages(p)}
+}
+func (a *Agent) ImageInput() *imageinput.Service { return a.imageInput.service }
+func (a *Agent) processToolImages(ctx context.Context, text string, images []string) imageResult {
 	if len(images) == 0 {
-		return text, nil
+		return imageResult{text: text}
 	}
-	if a.nativeImages {
-		return text, nil
+	if a.imageInput.native {
+		return imageResult{text: text}
 	}
-	summary, err := a.imageInput.Understand(ctx, a.modelRef, images, a.Session().Snapshot, a.svc.sink)
+	summary, err := a.imageInput.service.Understand(ctx, a.modelRef, images, a.Session().Snapshot, a.svc.sink)
 	if err != nil {
-		return text + fmt.Sprintf("\n[Image understanding unavailable: %v. The tool already executed; its text result remains valid. Do not claim to have seen the image or repeat the original action to retry image understanding.]", err), nil
+		return imageResult{text: text + fmt.Sprintf("\n[Image understanding unavailable: %v. The tool already executed; its text result remains valid. Do not claim to have seen the image or repeat the original action to retry image understanding.]", err), diagnostic: err}
 	}
-	return imageinput.AppendSummary(text, summary), summary
+	return imageResult{text: imageinput.AppendSummary(text, summary), summary: summary}
 }
 
 func supportsNativeImages(p provider.Provider) bool {
 	info, ok := p.(provider.ModelInfoProvider)
 	return ok && info.ModelInfo().SupportsInput(provider.ModalityImage)
-}
-
-// outcomeRunState preserves explicit execution evidence before image enrichment.
-// The fallback covers older and synthetic outcomes that have no image prepass.
-func outcomeRunState(o toolOutcome) provider.ToolRunState {
-	if o.runState != "" {
-		return o.runState
-	}
-	if !o.executed {
-		return provider.ToolRunNotStarted
-	}
-	return provider.ToolResultRunState(provider.Message{Content: o.output + "\n" + o.errMsg})
 }
