+++ internal/agent/usecapability_test.go
@@ -2,7 +2,6 @@ package agent
 
 import (
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -19,7 +18,6 @@ import (
 	"reasonix/internal/config"
 	"reasonix/internal/event"
 	"reasonix/internal/evidence"
-	"reasonix/internal/imageinput"
 	"reasonix/internal/mcplaunch"
 	"reasonix/internal/permission"
 	"reasonix/internal/plugin"
@@ -332,57 +330,6 @@ func imageMCPServer(t *testing.T, toolCalls *atomic.Int32, payload string) *http
 		w.Header().Set("Content-Type", "application/json")
 		_ = json.NewEncoder(w).Encode(map[string]any{"jsonrpc": "2.0", "id": *request.ID, "result": result})
 	}))
-}
-
-func TestPlannerFirstOnDemandMCPCallPreservesImages(t *testing.T) {
-	t.Setenv("REASONIX_CACHE_HOME", t.TempDir())
-	payload := base64.StdEncoding.EncodeToString([]byte("png-bytes"))
-	var toolCalls atomic.Int32
-	server := imageMCPServer(t, &toolCalls, payload)
-	defer server.Close()
-
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	host := plugin.NewHost()
-	defer host.Close()
-	spec := plugin.Spec{Name: "image", Type: "http", URL: server.URL, Authorized: true}
-	runtime := NewMCPCapabilityRuntime(ctx, host, []plugin.Spec{spec}, tool.NewRegistry(), nil)
-	proxy := runtime.NewFrontend(capability.NewLedger(), nil)
-	reg := tool.NewRegistry()
-	reg.Add(proxy)
-	prov := &scriptedProvider{name: "p", turns: [][]provider.Chunk{
-		{toolCallChunk("image-call", "use_capability", `{"action":"call","capability_id":"mcp-tool:image/screenshot","arguments":{}}`), {Type: provider.ChunkDone}},
-		{{Type: provider.ChunkText, Text: "done"}, {Type: provider.ChunkDone}},
-	}}
-	session := NewSession("sys")
-	vision := &summaryProvider{}
-	planner := NewPlannerAgent(prov, reg, session, Options{ImageInput: &imageinput.Config{Model: "vision/model", Resolve: func(string) (provider.Provider, error) { return vision, nil }}}, event.Discard)
-	if host.HasClient("image") {
-		t.Fatal("test requires the MCP server to start on first tool dispatch")
-	}
-	if err := planner.Run(withNoClosedLoop(ctx), "take a screenshot"); err != nil {
-		t.Fatalf("Run: %v", err)
-	}
-	if got := toolCalls.Load(); got != 1 {
-		t.Fatalf("image tools/call count = %d, want 1", got)
-	}
-	wantImage := "data:image/png;base64," + payload
-	for _, message := range session.Messages {
-		if message.Role != provider.RoleTool || message.ToolCallID != "image-call" {
-			continue
-		}
-		if len(message.Images) != 1 || message.Images[0] != wantImage {
-			t.Fatalf("first on-demand MCP images = %v, want %q", message.Images, wantImage)
-		}
-		if !strings.Contains(message.Content, "captured [image: image/png]") {
-			t.Fatalf("first on-demand MCP text = %q, want image placeholder", message.Content)
-		}
-		if message.VisionSummary == nil || !strings.Contains(message.Content, "OCR: Z7") || vision.calls.Load() != 1 {
-			t.Fatal("on-demand image did not use summary service")
-		}
-		return
-	}
-	t.Fatal("no tool message recorded for first on-demand MCP call")
 }
 
 func blockingReaderMCPServer(t *testing.T, callStarted chan<- struct{}, releaseCall <-chan struct{}, toolCalls *atomic.Int32) *httptest.Server {
