+++ internal/provider/responses/responses.go
@@ -365,111 +365,6 @@ func splitInstructions(messages []provider.Message) (string, []provider.Message)
 	return messages[0].Content, messages[1:]
 }
 
-func messagesToInput(messages []provider.Message, vision, replayWebSearchItems, summary bool) []map[string]any {
-	input := make([]map[string]any, 0, len(messages)*2)
-	// Keep function outputs together before appending a vision user message,
-	// matching the Chat adapter while retaining provider-specific exclusions.
-	var pendingImages []map[string]string
-	flushImages := func() {
-		if len(pendingImages) > 0 {
-			parts := []map[string]string{{"type": "input_text", "text": "Images returned by the preceding tool call(s):"}}
-			parts = append(parts, pendingImages...)
-			input = append(input, map[string]any{"role": "user", "content": parts})
-			pendingImages = nil
-		}
-	}
-	for _, message := range messages {
-		if message.Role != provider.RoleTool {
-			flushImages()
-		}
-		switch message.Role {
-		case provider.RoleSystem, provider.RoleUser:
-			// Text-only turns keep the documented TextInput string shape.
-			// Vision-capable user turns with attached images switch to the
-			// InputItemList array form ({type:input_text} + {type:input_image})
-			// so the text and every image ride the same message, matching the
-			// MiMo/DashScope multimodal example. The system message is always
-			// plain text: images only attach to user turns.
-			if vision && message.Role == provider.RoleUser && len(message.Images) > 0 {
-				parts := make([]map[string]string, 0, len(message.Images)+1)
-				if message.Content != "" {
-					parts = append(parts, map[string]string{"type": "input_text", "text": message.Content})
-				}
-				for _, ref := range message.Images {
-					if part := inputImagePart(ref); part != nil {
-						parts = append(parts, part)
-					}
-				}
-				if len(parts) == 0 || (len(parts) == 1 && parts[0]["type"] == "input_text") {
-					input = append(input, map[string]any{"role": "user", "content": message.Content})
-				} else {
-					input = append(input, map[string]any{"role": "user", "content": parts})
-				}
-			} else {
-				input = append(input, map[string]any{"role": string(message.Role), "content": message.Content})
-			}
-		case provider.RoleAssistant:
-			var rawReasoning bool
-			input, rawReasoning = appendReasoningItems(input, message.ResponsesItems)
-			if !rawReasoning && message.ReasoningContent != "" {
-				// Reasoning items: the OpenAI base format only needs
-				// `content`. DashScope additionally requires a `summary`
-				// list ("Invalid 'summary': summary is required and must be
-				// a list for reasoning."). Other vendors (MiMo) do not
-				// define summary in their schema; sending it leaks the
-				// reasoning text into an extra field the server may echo
-				// back into the model context, doubling chain-of-thought
-				// each turn — so only send it where the wire demands it.
-				item := map[string]any{
-					"type":    "reasoning",
-					"content": []map[string]string{{"type": "reasoning_text", "text": message.ReasoningContent}},
-				}
-				if message.ReasoningID != "" {
-					// OpenAI Responses schema marks Reasoning.id required;
-					// round-trip the provider-issued id when we captured one.
-					item["id"] = message.ReasoningID
-				}
-				if message.ReasoningStatus != "" {
-					item["status"] = message.ReasoningStatus
-				}
-				if summary {
-					item["summary"] = []map[string]string{{"type": "summary_text", "text": message.ReasoningContent}}
-				}
-				input = append(input, item)
-			}
-			if replayWebSearchItems {
-				for _, raw := range message.ResponsesItems {
-					if item, ok := decodeReplayableWebSearchItem(raw); ok {
-						input = append(input, item)
-					}
-				}
-			}
-			if message.Content != "" || len(message.ToolCalls) == 0 {
-				input = append(input, map[string]any{"role": "assistant", "content": message.Content})
-			}
-			for _, call := range message.ToolCalls {
-				input = append(input, map[string]any{
-					"type": "function_call", "call_id": call.ID,
-					"name": call.Name, "arguments": call.Arguments,
-				})
-			}
-		case provider.RoleTool:
-			input = append(input, map[string]any{
-				"type": "function_call_output", "call_id": message.ToolCallID, "output": message.Content,
-			})
-			if vision {
-				for _, ref := range message.Images {
-					if part := inputImagePart(ref); part != nil {
-						pendingImages = append(pendingImages, part)
-					}
-				}
-			}
-		}
-	}
-	flushImages()
-	return input
-}
-
 func decodeReplayableWebSearchItem(raw json.RawMessage) (map[string]any, bool) {
 	if len(raw) == 0 || len(raw) > maxReplayableSearchItemBytes || !json.Valid(raw) {
 		return nil, false
